address CR issues
